% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplet_detection.R
\name{cm_null_estimate}
\alias{cm_null_estimate}
\title{Estimate the null distribution using central matching}
\usage{
cm_null_estimate(zz, bre = 120, df = 7, pct0 = c(0.2, 0.6), nulltype)
}
\arguments{
\item{zz}{Numeric vector. The dataset to be analyzed.}

\item{bre}{Integer. Number of breaks for histogram binning.}

\item{df}{Integer. Degrees of freedom for natural splines in Poisson regression.}

\item{pct0}{Numeric vector of length 2. Defines the lower and upper quantile thresholds for selecting data used in null estimation.}

\item{nulltype}{Integer. The type of null estimation model:
- `2`: Symmetric model using quadratic fitting.
- `3`: Asymmetric model allowing different variance estimates for left and right tails.}
}
\value{
A list containing:
  - `delta.hat`: Estimated mode of the null distribution (for nulltype = 2).
  - `sigma.hat`: Estimated standard deviation of the null (for nulltype = 2).
  - `sigma.left`, `sigma.right`: Estimated left and right standard deviations (for nulltype = 3).
  - `p0`: Estimated null proportion for normalization.
}
\description{
This function estimates the null distribution of a dataset by performing truncation,
Poisson regression, and either a symmetric or asymmetric quadratic approximation.
}
