% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_preprocessing_pipeline}
\alias{run_preprocessing_pipeline}
\title{Run the Python Preprocessing Pipeline}
\usage{
run_preprocessing_pipeline(
  filtered_h5,
  raw_h5,
  atac_fragments,
  human_autosomes,
  output_directory,
  py_dir,
  rfilter_file = ""
)
}
\arguments{
\item{filtered_h5}{Path to the filtered_feature_bc_matrix.h5 file.}

\item{raw_h5}{Path to the raw_feature_bc_matrix.h5 file.}

\item{atac_fragments}{Path to the atac_fragments.tsv.gz file.}

\item{human_autosomes}{Path to the human_autosomes.txt file.}

\item{output_directory}{Directory where all outputs will be saved.}

\item{py_dir}{Directory where the Python scripts are located.}

\item{rfilter_file}{Optional path to a repetitive regions filter file.}
}
\value{
The path to the final output file ("colsum.csv") generated by the Python pipeline.
}
\description{
This function runs the three Python preprocessing steps:
1. Generate the single-cell barcode file.
2. Run the fragment overlap counter.
3. Process the overlap counter output.
}
